RF: move setup_parser() to main.py
To the one real consumer, and sets up cleanup work to avoid the circular
import problems Onyo has.
